CDD: Direct boot and FBE requirements
Android N provide support for filebase encryption, allowing files to be
encrypted with seperate keys bound to either the device or users'
credentials. This allows system processe that do not handle sensitive
user data (telephony, alarms, etc) to start before the user enters the
credentials and elimiate the double boot necessary for full disk
encryption.

This requires the following changes and afforances in the CDD:
- Sufficiently performant devices, with lockscreens, must use
  either FBE or FDE.
- Added Direct Boot Requirements
-- All Device must implement Direct Boot, regardless of encryption.
- Added FBE Requirements
-- DE anf CE keys must be bound to HW keystore and hardware
   root of trust (VB).
-- Must not be able to disable "secure startup" option on FBE
   devices. (In earlier versions of android the FDE implementation
   supported a "secure startup" option which required the user to
   provide their credentials before the device could boot. This option
   was disabled by default. FBE and Direct Boot provides a better
   solution and device implementations MUST NOT offer any method to
   unlock the CE protected storage without the user supplied
   credentials.)
-- MUST Support AES encryption as implemented in AOSP, MAY support
   others but AOSP MUST be used be default.
-- SHOULD make essential preloaded app directBootAware.

FDE requirements remain semantically unchanged, except it is not
required if the device implementaion use FBE.

Updated 3_10_accessibility to require that any pre-installed
accessibilty service MUST be direct boot aware on FBE devices.

BUG: 25897972
BUG: 27207717

Change-Id: I36fbce4937ebc161b09fdcb507db44f7b8990a3e