CDD: Add native APIs and require memory req. for cameras.
Android 7.0 adds support to control the camera through the native
APIs through the camera2 native APIs and we enlist the camera2 library
in the standard library list.

We also make it a requirement to have memory of at least 3 times the
maximum camera resolution, which was previously recommendation only.
See test android.hardware.camera2.StillCaptureTest#testAllocateBitmap
for details.

Bug: 28091251

Change-Id: I4cc94f2289ad8e759955697aa521a7aeb9c2bb6e