CDD: Clarify external cameras MAY be connected through non-USB ports.
Previously the CDD requirements for external camera assumed the device
is connected with the external camera through the USB port.

As this does not have to be necessarily the case, the requirements are
rewritten to allow device implementations to introduce new mechanisms
to connect the external camera(s) with the device.

Bug: 27930643

Change-Id: I9094b6d038d9194531fc7ca083f478fbf0dfca27