CDD: High perf VR devices requires BT 4.2 and LE Data Length
High performance VR device implementations are expected to be
used in a full-immersive nature, which means there would be a
need to have controllers typically external to the device.
Bluetooth 4.2 and LE Data Length extensions would be the baseline
foundation to grow an interoperable ecosystem of such accessories
upon which apps can rely on.

Bug: 26671343

Change-Id: I60d72f5a0746057bb2b63a4afe81f5e843cd45e9