CDD: Add requirements related to DND rules and policy set by apps
Android 7.0 allow apps to add rules that would automatically
enable or disable the Do-not-disturb (DND) mode on a certain
condition. For example an app capable of detecting activities
can turn on DND when the user is actually asleep at night.

Also, Android 7.0 introduces an API that allow apps to set the
DND notification policy  to suppress visual effects. 

Bug: 27953118

Change-Id: I57386eec34e46eb1b76e89372f1df975a8fec309