CDD: Introduce Safe Mode Requirements
Safe Mode, enabling users to boot into a state where only preinstalled
system apps are allowed to run, empowers the Android device users to
uninstall third-party apps.

The support of this mode is now STRONGLY RECOMMENDED as this mode can
be used to address cases where third-party apps might be interfering
with the user's capability to uninstall such apps.

Bug: 27337663

Change-Id: Ib921dc3ef7cca6db68d22e23d2063fdfb2877586