CDD: Require consistent system-wide root CAs across all Android
Android 7.0 is supporting the use case of apps to be configured with
app-specific root Certificate Authority (CA). Hence, now the policy
on the preinstalled root certificates in the system-trusted CA store
are more strictly enforced to make it harder to undermine the security
of the data communication from Android device implementations.

The guideline to handle public certificates are as below.

- Deprecated public CAs: MUST NOT be added.
- New public CAs not yet in AOSP: wait these public CAs to complete the
  Mozilla CA inclusion process and then file a feature request against
  Android (https://code.google.com/p/android/issues/entry to include the
  new public root CA to AOSP.
- private CAs that may be needed to securely access application servers
  or MNO(carrier) infrastructure, see:
  https://developer.android.com/training/articles/security-config.html

Bug: 18335321

Change-Id: I49bbc894c700d70d8049f9535550547fe1fce8e1