CDD: Update app private storage requirements.
Bumping the generic requirement up from 1.5GB to 3GB since:

a. the size of apps being distributed are growing
b. the statistics on median usage of the data/ partition is now over
2GB across devices from various geographies
c. the change was already hinted since 2 years ago

Reducing television requirements from 5GB to 4GB, because the
adoptable storage feature introduced in Android 6.0 are widely
adopted by televisions and can be used to meet the need for larger
app private storage

Bug: 26744064
Bug: 30684041

Change-Id: Ia9721380191093163412ea942cbb50bf40a32b38