CDD: Remove detail from Raw image decoder requirement
The AOSP provides an implementation of the Raw decoder in Skia
which will display the embedded JPEG (aka "preview") data --
if available in the image file -- or the DNG format data.

The change here removes this implementation detail from the
requirement, as the device implementations may have different
choices for how the Raw image file is decoded.

Bug: 27152493

Change-Id: I660198b12374f7050cfb49fdf11408589750ad5c