CDD: Require system privileged permissions to only be granted to apps      pre-installed in the whitelisted path.
The system permissions should not be extended to any app, just because
it's part of the system image but restricted to apps that are planned
to be part of the system. The API name change in Android 6.0 from
PROTECTION_FLAG_SYSTEM to PROTECTION_FLAG_PRIVILEGED further
adds to this point.

BUG: 33111571
Change-Id: Ibee24f8e424dc844e8cb49d5a7a0b56c3e3801aa