CDD: Clarify that the AOSP implementation of TEE is a preferred reference Section: 7.3.10. Fingerprint Sensor
The following requirement, has occassionaly been read that the use of
the AOSP implementation of the TEE is mandated.

"MUST prevent adding a fingerprint without first establishing a chain
of trust by having the user confirm existing or add a new device
credential (PIN/pattern/password) using the TEE as implemented in the
Android Open Source project"

Clarifed that there is an implementation of the TEE in the AOSP that
meets the requirement, and not that it is the only allowed
implementation.

Bug: 25121419
Change-Id: I19bf5535b7a532e52975fe99fceab5706be9f74b
(cherry picked from commit 32269fc2a910ea40d73fc4d67e929cfc7e6ca1fe)