CDD: Requirement for retention duration of user selection history.
Android 8.0 introduces the Smart sharing API, it learns about users'
personalized sharing preferences and better understands for each type of
content which are the right apps to share with. To support this API,
device implementations MUST keep a reasonable retention length of users'
sharing histories. It is Strongly Recommended to use the default
retention length.

bug: 33423136

Test: skipped.

Change-Id: I94bc1278aa2bfd11dce728e96bba61aa380d139b