CDD: Latency req. added for Pro audio when 3.5mm audio jack is omitted
We see some device makers omitting the 3.5mm audio jack, the interface
we have previously used to define the audio latency requirement for the
Pro Audio feature.

Hence we're updating the requirement to clarify that the audio latency
in this case should be measured on the USB host port interface.

Test: N/A
Bug: 35419299
Change-Id: I10d64ec7b9b4375d25b9438a583d87fb9963f70f