CDD: Require a default passcode to wrap the encryption key
If the user has not specified a lock screen credential, the process for
recovering the disk encryption key should still be bound to Keymaster
and the root of trust, so that an attacker who changes the OS to an
unsigned OS can't easily recover the disk encryption key. A default
passcode is the easy way to achieve that.

Given this, we are changing "SHOULD" to "MUST".

Bug: 33744049
Change-Id: I8e5026f394a8e4e6902f2b86449b367b6668f13b