CDD: Recommend simultaneous decoding of secure streams on Android Television devices.
Video content providers would like to have 2 secure videos decoded
to improve the transition between titles video playback.
This can't be done today because most Android devices can only
instantiate a single secure decoder at a time.

The industry is not yet ready to adopt this universally but the
requirement is added in Android 8.0 as a STRONGLY RECOMMENDED clause
to signal the intention to introduce a more strict requirement in the
future.

Bug: 25978754
Test: N/A
Change-Id: Ide72f348dbf89b712d561a7cfd4bfb1382150cd6