CDD: RECOMMEND against the use of Audio Adapter Accessory Mode on USB Type-C devices.
The Analog Adapter Accessory Mode as described in the USB Type-C
specification appendix A multiplex analog audio signals onto the
USB Type-C pins. The support of this feature might create confusion
when the support is not universal across different Android devices
with USB Type-C ports that for users would be indistinguishable,
and could create a loophole against the key event mapping
requirements for analog audio ports described in the
CDD section 7.8.2.1 Analog Audio Ports.

Bug: 35419794
Test: N/A
Change-Id: I0f449b2b6a980133c7bc1fd5fc2904f0d21b3737